+++ src/main/java/hu/vizicsaba/betvictortask/textstorageservice/service/model/TextStorageResponse.java
@@ -13,7 +13,7 @@ public class TextStorageResponse {
 
     private String mostFrequentWord;
 
-    private Integer averageParagraphSize;
+    private Double averageParagraphSize;
 
     private Double averageParagraphProcessingTime;
